Added commutation calibration and commit to flash MAVLink parameters

+++ Headers/hardware/HWSpecific.h
@@ -77,7 +77,7 @@ static const int EncoderSignMap[AXIS_CNT] = {
 };
 #endif
 
-extern int AxisHomePositions[AXIS_CNT];
+extern float AxisHomePositions[AXIS_CNT];
 extern float AxisCalibrationSlopes[AXIS_CNT];
 extern float AxisCalibrationIntercepts[AXIS_CNT];
 extern float AxisTorqueLoopKp[AXIS_CNT];
