+++ Source/hardware/HWSpecific.c
@@ -10,10 +10,10 @@
 // All of these get initialized over CAN at boot time (from parameters stored in flash on the AZ axis),
 // so they all get initialized to 0 here
 
-int AxisHomePositions[AXIS_CNT] = {
-    0,
-    0,
-    0
+float AxisHomePositions[AXIS_CNT] = {
+    0.0,
+    0.0,
+    0.0
 };
 
 float AxisCalibrationSlopes[AXIS_CNT] = {
